Preload design thumbs for a nicer UX
