+++ client/landing/gutenboarding/available-designs.ts
@@ -1,3 +1,14 @@
+/**
+ * External dependencies
+ */
+import { addQueryArgs } from '@wordpress/url';
+
+/**
+ * Internal dependencies
+ */
+import { isEnabled } from '../../config';
+import type { Design } from './stores/onboard/types';
+
 const availableDesigns: Readonly< AvailableDesigns > = {
 	featured: [
 		{
@@ -135,5 +146,38 @@ const availableDesigns: Readonly< AvailableDesigns > = {
 
 export default availableDesigns;
 interface AvailableDesigns {
-	featured: Array< import('./stores/onboard/types').Design >;
+	featured: Design[];
+}
+
+export const getDesignImageUrl = ( design: Design ) => {
+	// We temporarily show pre-generated screenshots until we can generate tall versions dynamically using mshots.
+	// See `bin/generate-gutenboarding-design-thumbnails.js` for generating screenshots.
+	// https://github.com/Automattic/mShots/issues/16
+	// https://github.com/Automattic/wp-calypso/issues/40564
+	if ( ! isEnabled( 'gutenboarding/mshot-preview' ) ) {
+		return `/calypso/page-templates/design-screenshots/${ design.slug }_${ design.template }_${ design.theme }.jpg`;
+	}
+
+	const mshotsUrl = 'https://s.wordpress.com/mshots/v1/';
+	const previewUrl = addQueryArgs( design.src, {
+		font_headings: design.fonts.headings,
+		font_base: design.fonts.base,
+	} );
+	return mshotsUrl + encodeURIComponent( previewUrl );
+};
+
+/**
+ * Asynchronously load available design images
+ */
+export function preloadDesignThumbs() {
+	if ( typeof window !== 'undefined' ) {
+		availableDesigns.featured.forEach( ( design: Design ) => {
+			const href = getDesignImageUrl( design );
+			const link = document.createElement( 'link' );
+			link.rel = 'preload';
+			link.as = 'image';
+			link.href = href;
+			document.head.appendChild( link );
+		} );
+	}
 }
